Refactor of chaincode approve
Simplify Approve() function signature, similar to Install() and QueryInstalled().

Use fabric-gateway to simplify implementation and avoid need to deal with orderers when submitting the transaction. This both simplifies the client application (as it doesn't need to worry about orderers or gRPC connections to them) and avoid potential problems when BFT ordering is introduced (which will require the transaction to be submitted to multiple orderers). The Gateway service takes care of the submit details.

Refactor e2e test to remove duplication of parallel invocation code.

Better unit testing of context. Does not rely in exactly the same context being passed to gRPC calls; a child context will still pass.

Signed-off-by: Mark S. Lewis <mark_lewis@uk.ibm.com>