Simplify protobuf assignment in unit tests
Since protobuf messages cannot be shallow copied, serialization was previously used to assign the state of an expected protobuf to a gRPC service response parameter when mocking gRPC service invocations. This can be replaced by using the `proto.Merge()` utility function, which achieves the same result without the overhead of serialization.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>