Fix Go build
The lint job failed as the Go version used was too old to run current
versions of staticcheck.

Additionally, this change sets the scheduled build and security
vulnerability scan to run weekly instead of daily, since the rate of
change in the project is low. Both are also configured to be runnable on
demand from the GitHub UI.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>