Categorise chaincode operations by peer and gateway (#194)
Previous implementation was standalone functions, with all requiring
both a gRPC connection and a client signing identity. This required:

- An underisably large number of prameters.
- Unwritten conventions on parameter ordering.
- Documentation comments to highlight when the connection should be to a
specific peer or to an organization gateway.
- Duplication of arguments across multiple functions for calling code
driving the chaincode lifecycle.

This implementation defines a Peer and Gateway struct that hold a gRPC
connection and a client identity. Each of these types defines methods
appropriate to their logical role to:

- Avoid the need to provide gRPC connection and client identity
arguments on every call.
- Provide calling code with logical context for what they are
interacting with.
- Allow IDE code completion to suggest appropriate methods for the
connection type.

For example, this pseudo-code flow:

    chaincode.Install(ctx, peerConnection, id, chaincodePackage)
    chaincode.Approve(ctx, gatewayConnection, id, chaincodeDefinition)
    chaincode.Commit(ctx, gatewayConnection, id, chaincodeDefinition)

becomes:

    peer.Install(ctx, chaincodePackage)
    gateway.Approve(ctx, chaincodeDefinition)
    gateway.Commit(ctx, chaincodeDefinition)

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>