Remove Node vulnerability scan
The Node implementation is not yet stabalized and depends on the
legacy Fabric Node SDK. The dependency tree now contains dependencies
that contain known vulnerabilities.

This change disables the the vulnerability scan for the Node
implementation so that the vulnerability status of the Go implementation
can be more clearly observed from scan results.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>